requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.42.2
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dj-rest-auth==7.0.1
Django==5.0.6
django-allauth==65.11.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-yasg==1.21.10
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.32.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.190
ultralytics-thop==2.0.16
uritemplate==4.2.0
urllib3==2.5.0
